adding PLP logo to both the login and register pages also incorparating media query to handle overflow in small devices
