+++ lib/pages/register.dart
@@ -8,6 +8,8 @@ class RegisterPage extends StatefulWidget {
 }
 
 class _RegisterPageState extends State<RegisterPage> {
+  late double deviceHeight, deviceWidth;
+
   // Controllers for email and password fields
   final TextEditingController emailController = TextEditingController();
   final TextEditingController passwordController = TextEditingController();
@@ -18,9 +20,12 @@ class _RegisterPageState extends State<RegisterPage> {
 
   @override
   Widget build(BuildContext context) {
+    deviceHeight = MediaQuery.of(context).size.height;
+    deviceWidth = MediaQuery.of(context).size.width;
     return Scaffold(
       appBar: AppBar(
-        title: Text('Register to PocketWallet', style: TextStyle(color: Colors.white)), // AppBar title
+        title: Text('Register to PocketWallet',
+            style: TextStyle(color: Colors.white)), // AppBar title
         backgroundColor: Colors.teal, // AppBar background color
       ),
       body: Padding(
@@ -28,23 +33,39 @@ class _RegisterPageState extends State<RegisterPage> {
         child: Form(
           key: _formKey, // Form key for validation
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.center, // Center the content vertically
+            mainAxisAlignment:
+                MainAxisAlignment.center, // Center the content vertically
             children: <Widget>[
+              SizedBox(
+                height: deviceHeight * 0.2,
+                width: deviceWidth * 0.2,
+                child: Image.asset('assets/images/logo.png'),
+              ),
               Text(
                 'Register',
-                style: Theme.of(context).textTheme.headlineMedium, // Uses a predefined text style from the app's theme
+                style: Theme.of(context)
+                    .textTheme
+                    .headlineMedium, // Uses a predefined text style from the app's theme
               ),
-              SizedBox(height: 20), // Space between title and email field
+              SizedBox(
+                  height: deviceHeight *
+                      0.03), // Space between title and email field
 
               // Email field
               TextFormField(
-                controller: emailController, // Connect the controller to the email input
+                controller:
+                    emailController, // Connect the controller to the email input
                 decoration: InputDecoration(
                   labelText: 'Email', // Placeholder text
-                  border: OutlineInputBorder(borderSide: BorderSide(color: Colors.black)), // Border style
-                  focusedBorder: OutlineInputBorder(borderSide: BorderSide(color: Colors.black)), // Focused border
+                  border: OutlineInputBorder(
+                      borderSide:
+                          BorderSide(color: Colors.black)), // Border style
+                  focusedBorder: OutlineInputBorder(
+                      borderSide:
+                          BorderSide(color: Colors.black)), // Focused border
                 ),
-                keyboardType: TextInputType.emailAddress, // Show email keyboard layout
+                keyboardType:
+                    TextInputType.emailAddress, // Show email keyboard layout
                 validator: (value) {
                   if (value == null || value.isEmpty) {
                     return 'Please enter your email'; // Error if the field is empty
@@ -55,16 +76,23 @@ class _RegisterPageState extends State<RegisterPage> {
                   return null; // No error if the field is valid
                 },
               ),
-              SizedBox(height: 10), // Space between email and password fields
+              SizedBox(
+                  height: deviceHeight *
+                      0.03), // Space between email and password fields
 
               // Password field
               TextFormField(
-                controller: passwordController, // Connect the controller to the password input
+                controller:
+                    passwordController, // Connect the controller to the password input
                 obscureText: true, // Hides the input for passwords
                 decoration: InputDecoration(
                   labelText: 'Password', // Placeholder text
-                  border: OutlineInputBorder(borderSide: BorderSide(color: Colors.black)), // Border style
-                  focusedBorder: OutlineInputBorder(borderSide: BorderSide(color: Colors.black)), // Focused border
+                  border: OutlineInputBorder(
+                      borderSide:
+                          BorderSide(color: Colors.black)), // Border style
+                  focusedBorder: OutlineInputBorder(
+                      borderSide:
+                          BorderSide(color: Colors.black)), // Focused border
                 ),
                 validator: (value) {
                   if (value == null || value.isEmpty) {
@@ -76,7 +104,9 @@ class _RegisterPageState extends State<RegisterPage> {
                   return null; // No error if the field is valid
                 },
               ),
-              SizedBox(height: 20), // Space before the register button
+              SizedBox(
+                  height:
+                      deviceHeight * 0.03), // Space before the register button
 
               // Register button
               ElevatedButton(
@@ -85,7 +115,8 @@ class _RegisterPageState extends State<RegisterPage> {
                   if (_formKey.currentState!.validate()) {
                     try {
                       // Create a new user with Firebase
-                      UserCredential userCredential = await _auth.createUserWithEmailAndPassword(
+                      UserCredential userCredential =
+                          await _auth.createUserWithEmailAndPassword(
                         email: emailController.text.trim(),
                         password: passwordController.text.trim(),
                       );
@@ -97,12 +128,14 @@ class _RegisterPageState extends State<RegisterPage> {
                       if (e.code == 'weak-password') {
                         errorMessage = 'The password provided is too weak.';
                       } else if (e.code == 'email-already-in-use') {
-                        errorMessage = 'The account already exists for that email.';
+                        errorMessage =
+                            'The account already exists for that email.';
                       } else {
                         errorMessage = 'An error occurred. Please try again.';
                       }
                       // Show error message
-                      ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text(errorMessage)));
+                      ScaffoldMessenger.of(context)
+                          .showSnackBar(SnackBar(content: Text(errorMessage)));
                     }
                   }
                 },
@@ -112,12 +145,14 @@ class _RegisterPageState extends State<RegisterPage> {
                 ),
                 child: Text('Register'), // Button text
               ),
-              SizedBox(height: 10), // Space before the login link
+              SizedBox(
+                  height: deviceHeight * 0.03), // Space before the login link
 
               // Login link
               TextButton(
                 onPressed: () {
-                  Navigator.pushNamed(context, '/'); // Navigate back to Login page
+                  Navigator.pushNamed(
+                      context, '/'); // Navigate back to Login page
                 }, // Login link text
                 style: TextButton.styleFrom(
                   foregroundColor: Colors.teal, // Text color
